Remove the Package and Program AST types.
These are above the level that has anything to do with syntax,
and don't even correspond to a single file.

Also removes the comment about what commit of ponyc this is based on.
The parser will continue to be based on a specific commit of ponyc,
but the AST definitions going forward will be just based on the parser.